redid SVE current version, kept new one commented out, moved all mod code to the mod file, learned that my commit mesages are public
