+++ StardewArchipelago/Constants/Modded/ModVersions.cs
@@ -30,7 +30,8 @@ namespace StardewArchipelago.Constants.Modded
             { ModNames.TRACTOR, $"4.21.{WILDCARD}" },
             //{ ModNames.WELLWICK, "1.0.0" },
             //{ ModNames.YOBA, "1.0.0" },
-            { ModNames.SVE, $"1.15.10" },
+            { ModNames.SVE, $"1.14.{WILDCARD}" },
+            //{ ModNames.SVE, $"1.15.10" }, less SVE versioning dirty edits 
             { ModNames.DISTANT_LANDS, $"2.2.{WILDCARD}" },
             { ModNames.LACEY, $"1.4.{WILDCARD}" },
             //{ ModNames.BOARDING_HOUSE, "4.0.16" },
@@ -53,8 +54,8 @@ namespace StardewArchipelago.Constants.Modded
             {
                 ModNames.SVE, new List<ContentPatcherRequirement>()
                 {
-                    //new(ModNames.AP_CP_SVE_PATCH, "2.1.x"),
-                    //new(ModNames.AP_CP_SVE, "1.14.46"),
+                    new(ModNames.AP_CP_SVE_PATCH, "2.1.x"),
+                    new(ModNames.AP_CP_SVE, "1.14.46"),
                 }
             },
         };
